Fixes converter station deps cycle
